+++ R/skimpy_ops.R
@@ -122,6 +122,16 @@
 }
 
 mvp_power_scalar <- function(S,n){
+
+  if(length(n)>1){
+    jj <- table(n)
+    out <- 0
+    for(i in seq_along(jj)){
+      out <- out + as.numeric(jj[i])*Recall(S,as.numeric(names(jj[i])))
+    }
+    return(out)
+  }
+
   stopifnot(n==round(n))
   if(n<0){
     stop("negative powers not implemented")